#2537 All scheduled tasks now extend SchedulerCommand, and they will output their execution time to InfluxDB -> Grafana
