+++ app/Console/Commands/Scheduler/Telemetry/Telemetry.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Commands\Scheduler\Telemetry;
 
+use App\Console\Commands\Scheduler\SchedulerCommand;
 use App\Console\Commands\Scheduler\Telemetry\Measurement\DungeonRouteCount;
 use App\Console\Commands\Scheduler\Telemetry\Measurement\MachineStats;
 use App\Console\Commands\Scheduler\Telemetry\Measurement\Measurement;
@@ -10,11 +11,10 @@ use App\Console\Commands\Scheduler\Telemetry\Measurement\QueueSize;
 use App\Console\Commands\Scheduler\Telemetry\Measurement\TeamCount;
 use App\Console\Commands\Scheduler\Telemetry\Measurement\UserCount;
 use Exception;
-use Illuminate\Console\Command;
 use InfluxDB\Database;
 use TrayLabs\InfluxDB\Facades\InfluxDB;
 
-class Telemetry extends Command
+class Telemetry extends SchedulerCommand
 {
     /**
      * The name and signature of the console command.
@@ -63,15 +63,17 @@ class Telemetry extends Command
      */
     public function handle(): int
     {
-        $points = [];
+        return $this->trackTime(function () {
+            $points = [];
 
-        foreach ($this->measurements as $measurement) {
-            $points = array_merge($points, $measurement->getPoints());
-        }
+            foreach ($this->measurements as $measurement) {
+                $points = array_merge($points, $measurement->getPoints());
+            }
 
-        // This function does actually exist
-        InfluxDB::writePoints($points, Database::PRECISION_SECONDS);
+            // This function does actually exist
+            InfluxDB::writePoints($points, Database::PRECISION_SECONDS);
 
-        return 0;
+            return 0;
+        });
     }
 }
